refactoring components
Added components folder

Added components to components folder
